projects/EduGuide-QuestionGen/requirements.txt
# requirements.txt - Python 依赖

# HTTP 客户端
httpx>=0.24.0

# 环境变量管理
python-dotenv>=1.0.0

# Streamlit 前端（可选）
streamlit>=1.28.0

# JSON 处理
pydantic>=2.0.0
